fix meals

+++ routes/api.php
@@ -38,16 +38,13 @@ Route::middleware(['auth:sanctum'])->group(function () {
       Route::delete('/profile/{id}', [ProfileController::class, 'destroy'])->name('profiles.destroy');
 });
 
-
-
-
 //meal
 Route::controller(MealController::class)->group(function(){
     Route::get('/meals','index')->middleware('auth:sanctum');
-    Route::post('/meal','store');
-    Route::get('/meal/{meal}','show')->middleware('admin');
-    Route::put('/meal/{meal}','update');
-    Route::delete('/meal/{meal}','destroy');
+    Route::post('/meal','store')->middleware(['auth:sanctum','admin']);
+    Route::get('/meal/{meal}','show')->middleware('auth:sanctum');
+    Route::put('/meal/{meal}','update')->middleware(['auth:sanctum','admin']);
+    Route::delete('/meal/{meal}','destroy')->middleware(['auth:sanctum','admin']);
 });
